refactor(reviews): extract to another path
also fix star selector and margin
